Fix test_ikfaskt_plugins.sh to run on Ubuntu >= 22.04

+++ moveit_kinematics/test/test_ikfast_plugins.sh
@@ -7,16 +7,18 @@
 
 set -e # fail script on error
 
-sudo update-alternatives --install /usr/bin/python python /usr/bin/python2 1
+if [ -f /usr/bin/python2 ]; then
+	sudo update-alternatives --install /usr/bin/python python /usr/bin/python2 1
+fi
 sudo update-alternatives --install /usr/bin/python python /usr/bin/python3 2
 sudo apt-get -q update
 
-if [ "$ROS_DISTRO" == "noetic" ]; then
-	sudo update-alternatives --set python /usr/bin/python3
-	sudo apt-get -qq install python3-lxml python3-yaml
-else
+if [ "$ROS_DISTRO" == "melodic" ]; then
 	sudo update-alternatives --set python /usr/bin/python2
 	sudo apt-get -qq install python-lxml python-yaml
+else
+	sudo update-alternatives --set python /usr/bin/python3
+	sudo apt-get -qq install python3-lxml python3-yaml
 fi
 
 # Clone moveit_resources for URDFs. They are not available before running docker.
